pyproject.toml
[tool.semantic_release]
version_variables = ["vmaas/common/constants.py:VMAAS_VERSION"]
version_toml = ["pyproject.toml:project.version"]
commit_author = "vmaas-bot <40663028+vmaas-bot@users.noreply.github.com>"

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "build",
    "chore",
    "ci",
    "docs",
    "feat",
    "fix",
    "perf",
    "style",
    "refactor",
    "test",
]
minor_tags = ["feat"]
patch_tags = ["build", "chore", "ci", "fix", "perf", "refactor"]

[project]
name = "vmaas"
version = "2.103.8"
readme = "README.md"
packages = []
package-mode = false
requires-python = "~3.12"

[tool.poetry.dependencies]
app-common-python = ">=0.2.5"
apscheduler = "^3.10.4"
boto3 = ">=1.36.2"
botocore = ">=1.36.2"
connexion = {extras = ["swagger-ui"],version = "^3.2.0"}
gitpython = ">=3.1.30"
iso8601 = ">=0.1.12"
natsort = ">=7.0.1"
psycopg2-binary = ">=2.8.6"
prometheus-client = ">=0.8.0"
pyopenssl = ">=25.0.0"
python-dateutil = ">=2.8.1"
pytz = ">=2020.4"
pyyaml = ">=5.4.1"
requests = ">=2.32.0"
schema = ">=0.7.3"
uvicorn = "^0.35.0"
watchtower = ">=1.0.0"
attrs = "^25.0.0"
zstandard = "^0.23.0"

[tool.poetry.group.dev.dependencies]
coverage = ">=4.5.3"
pur = ">=5.4.1"
pytest = ">=5.0.1"
pylint = ">=2.3.1"
pytest-cov = ">=2.7.1"
testing-postgresql = ">=1.3.0"
ipython = ">=8.18.1"
mypy = ">=1.8.0"
flake8 = "7.3.0"
types-psycopg2 = "*"
types-requests = "*"
types-python-dateutil = "*"
types-PyYAML = "*"
types-pyOpenSSL = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_any_generics = true
disallow_untyped_defs = true
disallow_untyped_calls = true
warn_return_any = true
warn_unreachable = true
follow_imports = "silent"
packages = """
    vmaas.reposcan.database.csaf_store,
    vmaas.reposcan.database.release_store,
    vmaas.reposcan.database.test.test_csaf,
    vmaas.reposcan.redhatcsaf,
    vmaas.reposcan.redhatrelease
"""
